+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.26.0]
+
+## Added
+- Support for the TCPS protocol for cloud databases in the Operations Insights service
+- Support for AIX hosts that are monitored via Enterprise Manager in the Operations Insights service
+- Support for lookups in the Logging Analytics service
+- Support for creating, updating, deleting and listing dashboards in the Management Dashboard Service
+- Support for changing the compartment of a dashboard in the Management Dashboard Service
+
+## Changed
+- Minimum required [OCI Python SDK](https://github.com/oracle/oci-python-sdk) changed to 2.105.0
+
 ## [4.25.0]
 
 ## Added